Convert nimbus.ts to nimbus.js with jsdoc type annotations
